Enable use of core.hooksPath
This is a far better option as we will not have to manually update each
repository when our hooks repository is updated.

The main disadvantage is a (relative) inability to tweak hooks depending
on the repository, but this will just force us to be more selective in
determining which scripts activate.

We still allow the use of init.templateDir to use hooks for users of
older versions of Git.